+++ src/main/java/com/hangout/hangout/domain/post/entity/PostHits.java
@@ -1,6 +1,6 @@
-package com.hangout.hangout.Post.domain.entity;
+package com.hangout.hangout.domain.post.entity;
 
-import com.hangout.hangout.user.domain.entity.User;
+import com.hangout.hangout.domain.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
