+++ src/main/java/com/hangout/hangout/domain/user/entity/Gender.java
@@ -1,0 +1,5 @@
+package com.hangout.hangout.domain.user.entity;
+
+public enum Gender {
+    MAN,WOMAN
+}
