+++ src/main/java/com/hangout/hangout/global/error/ResponseEntity.java
@@ -1,4 +1,4 @@
-package com.hangout.hangout.global.common.domain;
+package com.hangout.hangout.global.error;
 
 import lombok.AccessLevel;
 import lombok.Getter;
