--- src/main/java/com/hangout/hangout/user/domain/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.hangout.hangout.user.domain.entity;
-
-public enum Gender {
-    MAN,WOMAN
-}
